requirements.txt
# Core
python-dotenv>=1.0.0
requests>=2.31.0

# ArXiv API
arxiv>=2.1.0

# RSS Parsing
feedparser>=6.0.10

# News API (optional)
newsapi-python>=0.2.7

# Text Processing
beautifulsoup4>=4.12.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0

# Summarization (optional - for rule-based)
nltk>=3.8.0
sumy>=0.11.0

# LLM Integration (optional)
openai>=1.12.0
google-generativeai>=0.3.0

# Gmail API
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.118.0

# Scheduling
schedule>=1.2.0

# HTML Email
jinja2>=3.1.0
premailer>=3.10.0
